--- src/constants/index.js
@@ -1,7 +1,0 @@
-export const INCREMENT = "INCREMENT";
-export const DECREMENT = "DECREMENT";
-
-export const ADD_ITEM = "ADD_ITEM";
-export const REMOVE_ITEM = "REMOVE_ITEM";
-
-export const SET_MIN_RATING = "SET_MIN_RATING";
