+++ src/reducer/index.js
@@ -1,11 +1,11 @@
 import { combineReducers } from "redux";
-import counterReducer from "./counter";
-import order from "./order";
-import filters from "./filters";
-import restaurants from "./restaurants";
-import dishes from "./dishes";
-import reviews from "./reviews";
-import users from "./users";
+import counterReducer from "./counter/reducer";
+import order from "./order/reducer";
+import filters from "./filters/reducer";
+import restaurants from "./restaurants/reducer";
+import dishes from "./dishes/reducer";
+import reviews from "./reviews/reducer";
+import users from "./users/reducer";
 
 export default combineReducers({
   count: counterReducer,
