+++ src/reducer/order/constants.js
@@ -1,0 +1,2 @@
+export const ADD_ITEM = "ADD_ITEM";
+export const REMOVE_ITEM = "REMOVE_ITEM";
